memcached: Move all the event queuing/dequeing stuff to struct conn
Its not clear why this was ever in struct worker* in the first place,
because there's a rather obvious race that can happen. The processing
events is asynchronous and we have multiple connections, using a
single queue and assuming that we receive responses in the same
order that we sent requests over multiple connections seems like
a recipe for race conditions as there's no reason why ordering
of requests between different queues would be guaranteed by the event
system.