worker: Suppress clock skew issues leading to negative time differences
Its still not totally clear to me why this is happening, but one
reason may be clock skew issues that happen on HyperKit. What can happen
there is that time not always actually tick forward, during synchronization
with the host, it may skip forward or go backwards.

There's some more details here: https://engineering.docker.com/2019/02/addressing-time-drift-in-docker-desktop-for-mac/
but for now it doesn't seem safe to rely on the clock, so we just
truncate the delta to zero if it ends up being negative so that we
at least don't write to memory that we're not supposed to write to.